Binary release files from now on specify the toolchain
Retrieve it and provide it in the /dl/... API. Be backwards compatible by
providing a fallback for older releases that don't have the toolchain in
their filename.

Also set the `default` API parameter more precisely. On star releases the
default are installers, while on rakudo the default are the archive
releases.

Also in the /dl/... API skip properties that don't apply to the type of
file. `build_rev`, `backend`, `arch`, `toolchain` don't apply to non binary
files.