Remove the obvious installation methods
Noting that one can install Rakudo via e.g. the distro packages provides
no real benefit (as that's obvious) but clutters the presentation.

rakudo-pkg is kept, because it's standalone and wouldn't be seen at all
otherwise.

Alpine is kept because it's important to note that our precompiled packages
don't work there.